# src/agent/Hercules.WebApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Hercules.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/mesh/profiles/{name}": {
      "get": {
        "tags": [
          "MeshProfiles"
        ],
        "operationId": "GetMeshProfile",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/mesh/profiles/{name}/backends": {
      "get": {
        "tags": [
          "MeshProfiles"
        ],
        "operationId": "GetMeshProfileBackends",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/mesh/backend-status/{role}": {
      "get": {
        "tags": [
          "MeshProfiles"
        ],
        "operationId": "GetMeshBackendStatusByRole",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/quotas": {
      "get": {
        "tags": [
          "Quotas"
        ],
        "operationId": "QuotaStatus",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quotas/{scope}/{scopeId}": {
      "get": {
        "tags": [
          "Quotas"
        ],
        "operationId": "QuotaStatusByScope",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quotas/{scope}/{scopeId}/{type}": {
      "get": {
        "tags": [
          "Quotas"
        ],
        "operationId": "QuotaStatusByType",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/quotas/rate-limit": {
      "get": {
        "tags": [
          "Quotas"
        ],
        "operationId": "RateLimitInfo",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scopeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}